refactor: drawer content update
