+++ components/UI/Drawer/Drawer.tsx
@@ -1,100 +1,21 @@
 import useDisableZoom from "@/hooks/useDisableZoom";
 import { useMapClickHandlers } from "@/hooks/useMapClickHandlers";
 import { useWindowSize } from "@/hooks/useWindowSize";
-import { Data } from "@/mocks/TypesAreasEndpoint";
-import { dataFetcher } from "@/services/dataFetcher";
 import { useDrawerData, useIsDrawerOpen } from "@/stores/mapStore";
-import { dataTransformer } from "@/utils/dataTransformer";
-import { locationsURL } from "@/utils/urls";
-import {
-  Apple,
-  CopyAll,
-  DriveEta,
-  Google,
-  OpenInNew,
-} from "@mui/icons-material";
-import CloseIcon from "@mui/icons-material/Close";
-import { CircularProgress, Snackbar, TextField } from "@mui/material";
-import Box from "@mui/material/Box";
-import Button from "@mui/material/Button";
+import { Snackbar } from "@mui/material";
 import { default as MuiDrawer } from "@mui/material/Drawer";
-import formatcoords from "formatcoords";
 import {
   memo,
   MouseEvent,
-  ReactNode,
   useCallback,
   useEffect,
   useMemo,
   useState,
 } from "react";
-import useSWR from "swr";
-import GenericError from "../GenericError/GenericError";
 import styles from "./Drawer.module.css";
-import { getTimeAgo } from "@/utils/date";
-import FeedContent from "./components/FeedContent";
 import { useUrlPath } from "@/hooks/useUrlPath";
 import { useRouter } from "next/router";
-
-interface MapsButton {
-  label: string;
-  // eslint-disable-next-line no-unused-vars
-  urlCallback: (lat: number, lng: number) => void;
-  icon: ReactNode;
-  color: "primary" | "secondary" | "inherit";
-}
-
-export const generateGoogleMapsUrl = (lat: number, lng: number) => {
-  return `https://www.google.com/maps/?q=${lat},${lng}&ll=${lat},${lng}&z=21`;
-};
-
-export const generateAppleMapsUrl = (lat: number, lng: number) => {
-  return `http://maps.apple.com/?q=${lat},${lng}&ll=${lat},${lng}&z=18`;
-};
-
-export const openGoogleMapsUrl = (lat: number, lng: number) => {
-  window.open(generateGoogleMapsUrl(lat, lng), "_blank");
-};
-
-export const openAppleMapsUrl = (lat: number, lng: number) => {
-  window.open(generateAppleMapsUrl(lat, lng), "_blank");
-};
-
-export const openGoogleMapsDirectionUrl = (lat: number, lng: number) => {
-  window.open(
-    `https://www.google.com/maps?saddr=My+Location&daddr=${lat},${lng}`,
-    "_blank"
-  );
-};
-
-export const generateTweetUrl = (tweetId: string) => {
-  return `https://twitter.com/anyuser/status/${tweetId}`;
-};
-
-export const openTweetUrl = (tweetId: string) => {
-  window.open(generateTweetUrl(tweetId), "_blank");
-};
-
-export const mapsButtons: MapsButton[] = [
-  {
-    label: "Google Haritalarda Aç",
-    urlCallback: openGoogleMapsUrl,
-    icon: <Google className={styles.btnIcon} />,
-    color: "primary",
-  },
-  {
-    label: "Apple Haritalarda Aç",
-    urlCallback: openAppleMapsUrl,
-    icon: <Apple className={styles.btnIcon} />,
-    color: "inherit",
-  },
-  {
-    label: "Yol Tarifi Al",
-    urlCallback: openGoogleMapsDirectionUrl,
-    icon: <DriveEta className={styles.btnIcon} />,
-    color: "secondary",
-  },
-];
+import { Content } from "./components/Content";
 
 const Drawer = () => {
   useDisableZoom();
@@ -107,15 +28,6 @@ const Drawer = () => {
   const anchor = useMemo(
     () => (size.width > 768 ? "left" : "bottom"),
     [size.width]
-  );
-
-  const {
-    data: rawData,
-    isLoading,
-    error,
-  } = useSWR<Data | undefined>(
-    locationsURL(drawerData?.reference),
-    dataFetcher
   );
 
   function copyBillboard(url: string) {
@@ -133,152 +45,6 @@ const Drawer = () => {
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [isOpen]);
-
-  const list = () => {
-    const data = dataTransformer(rawData);
-
-    if (!drawerData) {
-      return null;
-    }
-
-    const formattedCoordinates = formatcoords([
-      drawerData.geometry.location.lat,
-      drawerData.geometry.location.lng,
-    ]).format();
-
-    const formattedTimeAgo = rawData && getTimeAgo(rawData.timestamp);
-    console.log(data);
-
-    const hasSource =
-      data.channel === "twitter" &&
-      // @ts-ignore: TODO tweet_id generic olmadığı için kızıyor, type ile fixlenebilir
-      data.extra_parameters.tweet_id &&
-      // @ts-ignore: TODO tweet_id generic olmadığı için kızıyor, type ile fixlenebilir
-      data.extra_parameters.tweet_id !== "";
-
-    return (
-      <Box
-        sx={{
-          width: size.width > 768 ? 400 : "full",
-          display: "flex",
-          height: "100%",
-          padding: "1rem 2rem 1rem 1rem",
-          flexDirection: "column",
-          overflow: "auto",
-        }}
-        role="presentation"
-        onKeyDown={(e) => toggler(e)}
-      >
-        {isLoading && (
-          <Box
-            sx={{
-              minHeight: "300px",
-              display: "flex",
-              justifyContent: "center",
-              alignItems: "center",
-            }}
-          >
-            <CircularProgress />
-          </Box>
-        )}
-        {error && <GenericError />}
-        {!isLoading && data && (
-          <div className={styles.content}>
-            <span className={styles.contentIdSection}>
-              ID: {drawerData.reference}
-            </span>
-            <h3 style={{ maxWidth: "45ch" }}>{data.formatted_address}</h3>
-            {formattedTimeAgo && (
-              <div className={styles.contentInfo}>
-                <svg viewBox="0 0 16 16" width="16" height="16" fill="#111111">
-                  <path d="M8.2 1.3c-3.7 0-6.7 3-6.7 6.7s3 6.7 6.7 6.7 6.7-3 6.6-6.7-3-6.7-6.6-6.7zM12 8.7h-4.5V4h1.3v3.3H12v1.4z" />
-                </svg>
-                <span>Bildirim zamanı: {formattedTimeAgo}</span>
-              </div>
-            )}
-            <div className={styles.contentInfo}>
-              <svg viewBox="0 0 16 16" width="16" height="16" fill="#111111">
-                <path d="M8 1A5.5 5.5 0 0 0 2.5 6.5a5.4 5.4 0 0 0 1.1 3.3s0.1 0.2 0.2 0.2L8 15l4.2-5c0 0 0.2-0.2 0.2-0.2l0 0A5.4 5.4 0 0 0 13.5 6.5 5.5 5.5 0 0 0 8 1Zm0 7.5a2 2 0 1 1 2-2 2 2 0 0 1-2 2Z" />
-                <path d="M8 6.5m-2 0a2 2 0 1 0 4 0 2 2 0 1 0-4 0" fill="none" />
-              </svg>
-              <span>{formattedCoordinates}</span>
-            </div>
-
-            <div className={styles.contentButtons}>
-              {mapsButtons.map((button) => (
-                <Button
-                  key={button.label}
-                  variant="contained"
-                  onClick={() => {
-                    button.urlCallback(
-                      drawerData.geometry.location.lat,
-                      drawerData.geometry.location.lng
-                    );
-                  }}
-                  color={button.color}
-                  className={styles.externalLinkButton}
-                  startIcon={button.icon}
-                >
-                  {button.label}
-                </Button>
-              ))}
-            </div>
-            <div>
-              <TextField
-                fullWidth
-                variant="standard"
-                size="small"
-                value={generateGoogleMapsUrl(
-                  drawerData.geometry.location.lat,
-                  drawerData.geometry.location.lng
-                )}
-                InputProps={{
-                  readOnly: true,
-                }}
-              />
-              <div className={styles.actionButtons}>
-                <Button
-                  variant="outlined"
-                  className={styles.clipboard}
-                  size="small"
-                  fullWidth
-                  onClick={() =>
-                    copyBillboard(
-                      `https://www.google.com/maps/@${drawerData.geometry.location.lat.toString()},${drawerData.geometry.location.lng.toString()},22z`
-                    )
-                  }
-                  startIcon={<CopyAll className={styles.btnIcon} />}
-                >
-                  Kopyala
-                </Button>
-                {hasSource && (
-                  <Button
-                    variant="outlined"
-                    className={styles.clipboard}
-                    fullWidth
-                    size="small"
-                    onClick={() =>
-                      window.open(
-                        // @ts-ignore: TODO tweet_id generic olmadığı için kızıyor, type ile fixlenebilir
-                        `https://twitter.com/anyuser/status/${data.extra_parameters?.tweet_id}`
-                      )
-                    }
-                    startIcon={<OpenInNew className={styles.btnIcon} />}
-                    color="secondary"
-                  >
-                    Kaynak
-                  </Button>
-                )}
-              </div>
-            </div>
-            <FeedContent content={data} />
-          </div>
-        )}
-
-        <CloseIcon onClick={(e) => toggler(e)} className={styles.closeButton} />
-      </Box>
-    );
-  };
 
   const handleClose = useCallback(
     (e: MouseEvent) => {
@@ -301,7 +67,9 @@ const Drawer = () => {
         open={isOpen}
         onClose={handleClose}
       >
-        {list()}
+        {!!drawerData && (
+          <Content drawerData={drawerData} onCopyBillboard={copyBillboard} />
+        )}
       </MuiDrawer>
     </div>
   );
